Moved the Amp Configurator class into its own file
