+++ www/configurator/src/app/hello/hello.js
