+++ www/configurator/src/app/hello/hello.spec.js
